23.07.09 11commits

+++ 003 Code/Backend/api/src/main/java/com/hanbat/zanbanzero/service/user/UserService.java
@@ -37,19 +37,25 @@ public class UserService implements UserDetailsService {
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     @Transactional
-    public void join(UserJoinDto dto) throws JsonProcessingException {
+    public User join(UserJoinDto dto) throws JsonProcessingException {
         dto.setEncodePassword(bCryptPasswordEncoder);
         User user = userRepository.save(User.of(dto));
         userMyPageRepository.save(UserMypage.createNewUserMyPage(user));
         userPolicyRepository.save(UserPolicy.createNewUserPolicy(user));
+        return user;
+    }
+
+    @Transactional
+    public UserInfoDto loginFromKeycloak(User u) throws JsonProcessingException {
+        User user = userRepository.findByUsername(u.getUsername());
+        if (user == null) user = join(UserJoinDto.of(u));
+        return UserInfoDto.of(user);
     }
 
     @Transactional
     public void withdraw(UserJoinDto dto) throws WrongRequestDetails {
         User user = userRepository.findByUsername(dto.getUsername());
-        if (bCryptPasswordEncoder.matches(dto.getPassword(), user.getPassword())) {
-            userRepository.delete(user);
-        }
+        if (bCryptPasswordEncoder.matches(dto.getPassword(), user.getPassword())) userRepository.delete(user);
         else throw new WrongRequestDetails("비밀번호 틀림");
     }
